fixing UT

+++ marshall_webapp/views/tests/test_views_transients_followup_obs.py
@@ -15,38 +15,40 @@ from sqlalchemy.orm import sessionmaker
 from fundamentals import tools
 from marshall_webapp.utKit import utKit
 from dryxPyramid.utKit import BaseTest
+home = expanduser("~")
 
-moduleDirectory = os.path.dirname(__file__)
+packageDirectory = utKit("").get_project_root()
+# settingsFile = packageDirectory + "/test_settings.yaml"
+settingsFile = home + "/git_repos/_misc_/settings/marshall/test_settings.yaml"
 
-su = tools(
-    arguments={"settingsFile": moduleDirectory + "/../../test_settings.yaml"},
-    docString=__doc__,
-    logLevel="DEBUG",
-    options_first=False,
-    projectName=None,
-    defaultSettingsFile=False
-)
-arguments, settings, log, dbConn = su.setup()
+exists = os.path.exists(settingsFile)
+if exists:
 
-# SETUP AND TEARDOWN FIXTURE FUNCTIONS FOR THE ENTIRE MODULE
-moduleDirectory = os.path.dirname(__file__)
-utKit = utKit(moduleDirectory)
-log, dbConn, pathToInputDir, pathToOutputDir = utKit.setupModule()
-utKit.tearDownModule()
+    su = tools(
+        arguments={"settingsFile": settingsFile},
+        docString=__doc__,
+        logLevel="DEBUG",
+        options_first=False,
+        projectName=None,
+        defaultSettingsFile=False
+    )
+    arguments, settings, log, dbConn = su.setup()
 
+    # SETUP PATHS TO COMMON DIRECTORIES FOR TEST DATA
+    moduleDirectory = os.path.dirname(__file__)
+    pathToInputDir = moduleDirectory + "/input/"
+    pathToOutputDir = moduleDirectory + "/output/"
 
-try:
-    shutil.rmtree(pathToOutputDir)
-except:
-    pass
-# COPY INPUT TO OUTPUT DIR
-shutil.copytree(pathToInputDir, pathToOutputDir)
+    try:
+        shutil.rmtree(pathToOutputDir)
+    except:
+        pass
+    # COPY INPUT TO OUTPUT DIR
+    shutil.copytree(pathToInputDir, pathToOutputDir)
 
-# Recursively create missing directories
-if not os.path.exists(pathToOutputDir):
-    os.makedirs(pathToOutputDir)
-
-# xt-setup-unit-testing-files-and-folders
+    # Recursively create missing directories
+    if not os.path.exists(pathToOutputDir):
+        os.makedirs(pathToOutputDir)
 
 
 class test_views_transients_followup_obs(BaseTest):
